# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	datafusion/core/src/datasource/file_format/avro.rs

+++ datafusion/core/tests/user_defined/user_defined_window_functions.rs
@@ -536,7 +536,7 @@ impl OddCounter {
         impl SimpleWindowUDF {
             fn new(test_state: Arc<TestState>) -> Self {
                 let signature =
-                    Signature::exact(vec![DataType::Float64], Volatility::Immutable);
+                    Signature::exact(vec![DataType::Int64], Volatility::Immutable);
                 Self {
                     signature,
                     test_state: test_state.into(),
